Move covariance_sv to netket.stats
