+++ Sources/Stats/py_stats.cc
@@ -3,8 +3,10 @@
 #include <iomanip>
 
 #include <pybind11/eigen.h>
+#include <pybind11/numpy.h>
 #include <pybind11/pybind11.h>
 
+#include "Sampler/vmc_sampling.hpp"
 #include "Stats/mc_stats.hpp"
 #include "Stats/obs_manager.hpp"
 #include "Utils/exceptions.hpp"
@@ -127,6 +129,65 @@ void AddStatsModule(py::module m) {
                     Rank-2 tensors should have shape `(N, M)` where `N` is the
                     number of samples in one Markov Chain and `M` is the number
                     of Markov Chains. Data should be in row major order.)EOF");
+
+  subm.def(
+      "covariance_sv",
+      [](py::array_t<Complex, py::array::c_style> s_values,
+         py::array_t<Complex, py::array::c_style> v_values) {
+        Eigen::Map<const VectorXcd> s_vector{s_values.data(), s_values.size()};
+
+        // Compute S -> S - E[S]
+        Complex mean = s_vector.mean();
+        MeanOnNodes(mean);
+
+        switch (s_values.ndim()) {
+          case 2:
+            NETKET_CHECK(v_values.ndim() == 3, InvalidInputError,
+                         "v_values has wrong dimension: " << v_values.ndim()
+                                                          << "; expected 3.");
+            return Gradient(
+                s_vector.array() - mean,
+                Eigen::Map<const RowMatrix<Complex>>{
+                    v_values.data(), v_values.shape(0) * v_values.shape(1),
+                    v_values.shape(2)});
+          case 1:
+            NETKET_CHECK(v_values.ndim() == 2, InvalidInputError,
+                         "v_values has wrong dimension: " << v_values.ndim()
+                                                          << "; expected 2.");
+            return Gradient(
+                s_vector.array() - mean,
+                Eigen::Map<const RowMatrix<Complex>>{
+                    v_values.data(), v_values.shape(0), v_values.shape(1)});
+          default:
+            NETKET_CHECK(false, InvalidInputError,
+                         "s_values has wrong dimension: "
+                             << s_values.ndim() << "; expected either 1 or 2.");
+        }  // end switch
+      },
+      py::arg{"s"}.noconvert(), py::arg{"v"}.noconvert(),
+      R"EOF(Computes the covariance of two random variables S, V from MCMC
+            data. Note that while S has to be a complex scalar variable,
+            V = (v[1], ..., v[m]) can be vector-valued. This is because this
+            function is designed to be used for computing the gradient of
+            expectation values <O>, where S is the local estimator of O and
+            V the vector of log-derivatives of the wavefunction.
+
+            This function estimates R = (r[1], ..., r[m]) where
+                r[i] = Cov[v[i], S] = 𝔼[conj(v[i]) * (S - 𝔼[S])].
+
+            Args:
+                s_values: A vector (or a matrix) of samples of S with shape
+                    `(N, M)` where `N` is the number of samples in every Markov
+                    Chain and `M` is the number of Markov Chains.
+                    A vector is considered to be an `(N, 1)` matrix.
+                v_values: A matrix (or a rank-3 array) of samples of V.
+                    Each row of the matrix must correspond to one component v[i].
+                    For a rank-3 array, its shape is `(N, M, m)` where `N`
+                    is the number of samples, `M` is the number of Markov Chains,
+                    and `m` is the dimension of V. A `(N, m)` matrix
+                    is treated an an `(N, 1, m)` array.
+
+        )EOF");
 }
 }  // namespace detail
 }  // namespace netket
